change range + change variable type on double

+++ PascalServer/Models/PascalJson.cs
@@ -9,13 +9,13 @@ namespace PascalServer.Models
     {
         public List<PascalRow> TriangleRows { get; private set; }
 
-        public PascalJson(long[,] pascalTriangleTable, int tableLength, int modulo = 0)
+        public PascalJson(double[,] pascalTriangleTable, int tableLength, int modulo = 0)
         {
             TriangleRows = new List<PascalRow>();
             BuildTriangle(pascalTriangleTable, tableLength, modulo);
         }
 
-        private void BuildTriangle(long[,] pascalTriangleTable, int tableLength, int modulo = 0)
+        private void BuildTriangle(double[,] pascalTriangleTable, int tableLength, int modulo = 0)
         {
             for (int i = 0; i < tableLength; i++)
             {
